refactor: update app/ import paths to features/ and shared/
