+++ app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { AuthForm } from "@/components/auth/AuthForm";
+import { AuthForm } from "@/features/auth/components/AuthForm";
 
 export default function LoginPage() {
   return <AuthForm mode="login" />;
