+++ app/auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import { AuthForm } from "@/components/auth/AuthForm";
+import { AuthForm } from "@/features/auth/components/AuthForm";
 
 export default function SignupPage() {
   return <AuthForm mode="signup" />;
